remove unused leaflet components, give "Maplibre" components more generic names, replace admin map
