+++ frontend/package.json
@@ -40,9 +40,6 @@
     "formik": "^2.4.2",
     "leaflet": "^1.9.4",
     "leaflet-control-geocoder": "^2.4.0",
-    "leaflet-measure": "^3.1.0",
-    "leaflet-side-by-side": "^2.2.0",
-    "leaflet.markercluster": "^1.5.3",
     "localforage": "^1.10.0",
     "maplibre-gl": "^4.7.1",
     "papaparse": "^5.4.1",
@@ -50,7 +47,6 @@
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.52.2",
     "react-leaflet": "^4.2.1",
-    "react-leaflet-cluster": "^2.1.0",
     "react-map-gl": "^7.1.7",
     "react-router-dom": "^6.14.2",
     "react-select": "^5.8.0",
