+++ frontend/src/components/maps/CompareMap/index.tsx
@@ -1,0 +1,3 @@
+import CompareMap from './CompareMap';
+
+export default CompareMap;
