+++ frontend/src/components/maps/DrawFieldMap/index.tsx
@@ -1,0 +1,3 @@
+import DrawFieldMap from './DrawFieldMap';
+
+export default DrawFieldMap;
