+++ frontend/src/components/maps/GeocoderControl.tsx
@@ -1,66 +1,77 @@
-import 'leaflet-control-geocoder/dist/Control.Geocoder.css';
-import L from 'leaflet';
-import Geocoder, { geocoders } from 'leaflet-control-geocoder';
-import { useEffect, useRef } from 'react';
-import { useMap } from 'react-leaflet/hooks';
+import '@maplibre/maplibre-gl-geocoder/dist/maplibre-gl-geocoder.css';
+import './GeocoderControl.css';
 
-export default function GeocoderControl({
-  position = 'topleft',
-  showMarker = false,
-}: {
-  position?: L.ControlPosition;
-  showMarker?: boolean;
-}) {
-  const map = useMap();
+import maplibregl from 'maplibre-gl';
+import { useEffect } from 'react';
+import { useMap } from 'react-map-gl/maplibre';
+import MaplibreGeocoder, {
+  CarmenGeojsonFeature,
+  MaplibreGeocoderApi,
+  MaplibreGeocoderFeatureResults,
+} from '@maplibre/maplibre-gl-geocoder';
 
-  const bboxPolyRef = useRef<L.Polygon | null>(null);
+export default function GeocoderControl() {
+  const { current: map } = useMap();
 
   useEffect(() => {
-    // Create geocoder control with default OSM/Nominatim geocoding provider
-    const GeocoderControl = new Geocoder({
-      defaultMarkGeocode: showMarker,
-      geocoder: new geocoders.Nominatim(),
-      position: position,
-    });
-    GeocoderControl.addTo(map);
+    if (!map) return;
 
-    // Create polygon for geocode bounding box and add to map
-    if (!showMarker) {
-      GeocoderControl.on('markgeocode', function (e) {
-        // Remove bounding box polygon if one exists
-        if (bboxPolyRef.current) {
-          map.removeLayer(bboxPolyRef.current);
+    const geocoderApi: MaplibreGeocoderApi = {
+      forwardGeocode: async (config) => {
+        const features: CarmenGeojsonFeature[] = [];
+        try {
+          const request = `https://nominatim.openstreetmap.org/search?q=${config.query}&format=geojson&polygon_geojson=1&addressdetails=1`;
+          const response = await fetch(request);
+          const geojson = await response.json();
+          geojson.features.forEach((feature, index) => {
+            const center = [
+              feature.bbox[0] + (feature.bbox[2] - feature.bbox[0]) / 2,
+              feature.bbox[1] + (feature.bbox[3] - feature.bbox[1]) / 2,
+            ];
+            const point: CarmenGeojsonFeature = {
+              id: `${index}`,
+              type: 'Feature',
+              geometry: {
+                type: 'Point',
+                coordinates: center,
+              },
+              place_name: feature.properties.display_name,
+              properties: feature.properties,
+              text: feature.properties.display_name,
+              place_type: ['place'],
+            };
+            features.push(point);
+          });
+        } catch (e) {
+          console.error(`Failed to forwardGeocode with error: ${e}`);
         }
-        // Create new bounding box polygon and add to map
-        const bbox = e.geocode.bbox;
-        const poly = L.polygon(
-          [
-            bbox.getSouthEast(),
-            bbox.getNorthEast(),
-            bbox.getNorthWest(),
-            bbox.getSouthWest(),
-          ],
-          { fill: false, color: '#f9ff33' }
-        ).addTo(map);
-        // Update state with new bounding box polygon
-        bboxPolyRef.current = poly;
-        // Zoom map to extent of bounding box polygon
-        map.fitBounds(poly.getBounds());
-        // Add event to remove polygon when clicked
-        // poly.addEventListener('click', () => {
-        //   if (confirm('Remove boundary?') && bboxPolyRef.current) {
-        //     map.removeLayer(bboxPolyRef.current);
-        //   }
-        // });
-      });
-    }
 
-    // Cleanup control after component unmounts or effect re-runs
+        return {
+          type: 'FeatureCollection',
+          features,
+        } as MaplibreGeocoderFeatureResults;
+      },
+      reverseGeocode: async () => {
+        // No-op, added to meet MaplibreGeocoderApi type requirements
+        return {
+          type: 'FeatureCollection',
+          features: [],
+        } as MaplibreGeocoderFeatureResults;
+      },
+    };
+
+    const geocoder = new MaplibreGeocoder(geocoderApi, {
+      collapsed: true,
+      debounceSearch: 800,
+      marker: false,
+      maplibregl: maplibregl,
+      showResultMarkers: true,
+    });
+    map.addControl(geocoder, 'top-left');
+
     return () => {
-      GeocoderControl.remove();
-
-      if (bboxPolyRef.current) {
-        map.removeLayer(bboxPolyRef.current);
+      if (map) {
+        map.removeControl(geocoder);
       }
     };
   }, [map]);
