+++ frontend/src/components/maps/MapViewMode.tsx
@@ -1,18 +1,18 @@
 import { useMapContext } from './MapContext';
-import MaplibreCompareMap from './MaplibreCompareMap';
-import MaplibreMap from './MaplibreMap';
+import CompareMap from './CompareMap';
+import HomeMap from './HomeMap';
 import PotreeViewer from './PotreeViewer';
 
-export default function MaplibreView() {
+export default function MapViewMode() {
   const { activeDataProduct, activeMapTool } = useMapContext();
 
   if (activeMapTool === 'compare') {
-    return <MaplibreCompareMap />;
+    return <CompareMap />;
   } else if (
     !activeDataProduct ||
     (activeDataProduct && activeDataProduct.data_type !== 'point_cloud')
   ) {
-    return <MaplibreMap />;
+    return <HomeMap />;
   } else {
     const copcPath = activeDataProduct.url;
     return <PotreeViewer copcPath={copcPath} />;
