+++ frontend/src/components/maps/layerProps.tsx
