+++ frontend/src/components/maps/styles/basemapStyles.ts
