+++ frontend/src/components/maps/styles/drawStyles.ts
