+++ frontend/src/components/pages/admin/DashboardMap.tsx
@@ -1,41 +1,80 @@
-import axios, { AxiosResponse } from 'axios';
-import { MapContainer } from 'react-leaflet/MapContainer';
-import { ScaleControl } from 'react-leaflet';
-import { ZoomControl } from 'react-leaflet/ZoomControl';
-import { useLoaderData } from 'react-router-dom';
+import 'maplibre-gl/dist/maplibre-gl.css';
+import { useState } from 'react';
+import Map, { NavigationControl, ScaleControl } from 'react-map-gl/maplibre';
 
-import ProjectMarkers from '../../maps/ProjectMarkers';
-import { Project } from '../projects/ProjectList';
-import MapLayersControl from '../../maps/MapLayersControl';
+import ProjectCluster from '../../maps/ProjectCluster';
+import ProjectPopup from '../../maps/ProjectPopup';
 
-export async function loader() {
-  const response: AxiosResponse<Project[]> = await axios.get(
-    `${import.meta.env.VITE_API_V1_STR}/projects?include_all=True`
-  );
-  if (response) {
-    return response.data;
-  } else {
-    return [];
-  }
-}
+import { PopupInfoProps } from '../../maps/HomeMap';
+import {
+  mapboxSatelliteBasemapStyle,
+  usgsImageryTopoBasemapStyle,
+} from '../../maps/styles/basemapStyles';
 
 export default function DashboardMap() {
-  const projects = useLoaderData() as Project[];
+  const [popupInfo, setPopupInfo] = useState<PopupInfoProps | null>(null);
+
+  const handleMapClick = (event) => {
+    const map: maplibregl.Map = event.target;
+
+    if (map.getLayer('unclustered-point')) {
+      const features = map.queryRenderedFeatures(event.point, {
+        layers: ['unclustered-point'],
+      });
+
+      if (features.length > 0) {
+        const clickedFeature = features[0];
+
+        if (clickedFeature.geometry.type === 'Point') {
+          const coordinates = clickedFeature.geometry.coordinates;
+
+          setPopupInfo({
+            feature: clickedFeature,
+            latitude: coordinates[1],
+            longitude: coordinates[0],
+          });
+        }
+      }
+    }
+  };
+
+  const handlePopupClose = () => setPopupInfo(null);
 
   return (
-    <MapContainer
-      center={[40.428655143949925, -86.9138040788386]}
-      preferCanvas={true}
-      zoom={8}
-      maxZoom={24}
-      scrollWheelZoom={true}
-      zoomControl={false}
-      worldCopyJump={true}
+    <Map
+      initialViewState={{
+        longitude: -86.9138040788386,
+        latitude: 40.428655143949925,
+        zoom: 8,
+      }}
+      style={{
+        width: '100%',
+        height: '100%',
+      }}
+      mapboxAccessToken={import.meta.env.VITE_MAPBOX_ACCESS_TOKEN || undefined}
+      mapStyle={
+        import.meta.env.VITE_MAPBOX_ACCESS_TOKEN
+          ? mapboxSatelliteBasemapStyle
+          : usgsImageryTopoBasemapStyle
+      }
+      reuseMaps={true}
+      onClick={handleMapClick}
     >
-      {projects && projects.length > 0 && <ProjectMarkers projects={projects} />}
-      <MapLayersControl />
-      <ZoomControl position="topleft" />
-      <ScaleControl position="bottomright" />
-    </MapContainer>
+      {/* Display marker cluster for all project centroids */}
+      <ProjectCluster includeAll={true} />
+
+      {/* Display popup when unclustered project marker clicked on */}
+      {popupInfo && (
+        <ProjectPopup
+          onClose={handlePopupClose}
+          popupInfo={popupInfo}
+          showActionButton={false}
+        />
+      )}
+
+      {/* General Controls */}
+      <NavigationControl />
+      <ScaleControl />
+    </Map>
   );
 }
